+++ src/Update/UpdateChecker.php
@@ -17,11 +17,7 @@ use Robo\Contract\ConfigAwareInterface;
  * Class UpdateChecker
  * @package Pantheon\Terminus\Update
  */
-class UpdateChecker implements
-    ConfigAwareInterface,
-    ContainerAwareInterface,
-    DataStoreAwareInterface,
-    LoggerAwareInterface
+class UpdateChecker implements ConfigAwareInterface, ContainerAwareInterface, DataStoreAwareInterface, LoggerAwareInterface
 {
     use ConfigAwareTrait;
     use ContainerAwareTrait;
@@ -29,9 +25,7 @@ class UpdateChecker implements
     use LoggerAwareTrait;
 
     const DEFAULT_COLOR = "\e[0m";
-    const INSTALLER_COMMAND =
-        'curl -O https://raw.githubusercontent.com/pantheon-systems/terminus-installer/master/builds/installer.phar';
-    const UPDATE_COMMAND = 'php installer.phar update';
+    const UPDATE_COMMAND = 'curl -O https://raw.githubusercontent.com/pantheon-systems/terminus-installer/master/builds/installer.phar && php installer.phar update';
     const UPDATE_NOTICE = <<<EOT
 A new Terminus version v{latest_version} is available.
 You are currently using version v{running_version}.