+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org)
 
+## 1.5.0 - 2017-08-17
 ### Changed
 - Updated the name of the `longname` field output by `upstream:list` to `label`. (#1747)
 - Updated the name of the `longname` field output by `upstream:info` to `label`. (#1747)